backend/requirements.txt
fastapi==0.98.0
uvicorn==0.22.0
pydantic==1.10.9
sqlalchemy==2.0.16
numpy==1.24.3
scipy==1.10.1
nltk==3.8.1
scikit-learn==1.2.2
transformers==4.30.2
python-multipart==0.0.6
pyttsx3==2.90
SpeechRecognition==3.10.0
PyAudio==0.2.13
whatsapp-api-client-python==0.0.40
jinja2==3.1.2
python-dotenv==1.0.0 